Task 1: Check 14

+++ 0x15-file_io/1-create_file.c
@@ -13,8 +13,8 @@
 int create_file(const char *filename, char *text_content)
 {
 int fd = open(filename, O_CREAT | O_WRONLY | O_TRUNC, 0600);
-ssize_t len = 0;
-int bytes_written = write(fd, text_content, len);
+size_t len = 0;
+ssize_t bytes_written = write(fd, text_content, len);
 
 if (filename == NULL)
 return (-1);
@@ -23,7 +23,7 @@ return (-1);
 if (text_content != NULL)
 {
 len = strlen(text_content);
-if (bytes_written == -1 || bytes_written != len)
+if (bytes_written == -1 || bytes_written != (ssize_t) len)
 {
 close(fd);
 return (-1);
